+++ ChangeLog.md
@@ -41,6 +41,7 @@ Table:
 - Adds several tests for Table APIs
 - Fixes issues for upsert and merge with etag matching
 - Allow any valid weak etag even though we know it will fail with a 412
+- Added check for table query validity
 
 ## 2021.12 Version 3.15.0